MWPW-147284: discard default params in merch link
that are generated via OST.
