+++ test/blocks/ost/ost.test.html.js
@@ -1,15 +1,62 @@
 import { expect } from '@esm-bundle/chai';
 
-import { mockOstDeps, unmockOstDeps } from './mocks/ost-utils.js';
+import { mockOstDeps } from './mocks/ost-utils.js';
+import { CheckoutWorkflow, CheckoutWorkflowStep } from '../../../libs/deps/commerce.js';
+import { DEFAULT_CTA_TEXT, createLinkMarkup } from '../../../libs/blocks/ost/ost.js';
 
-afterEach(() => {
-  unmockOstDeps();
-});
+const { perpM2M } = await fetch('./mocks/wcs-artifacts-mock.json').then((res) => res.json());
+const defaults = {
+  checkoutWorkflow: 'UCv3',
+  checkoutWorkflowStep: 'email',
+};
+const osi = 'cea462e983f649bca2293325c9894bdd';
+const promo = 'test-promo';
+const texts = {
+  buy: DEFAULT_CTA_TEXT,
+  try: 'free-trial',
+};
+const types = {
+  checkoutUrl: 'checkoutUrl',
+  price: 'price',
+  opticalPrice: 'opticalPrice',
+};
 
-describe('loadOstEnv', async () => {
+function assertLink(link, offer, params, text = texts.buy) {
+  const { searchParams } = new URL(link.href);
+  Object.entries(params).forEach(([key, value]) => {
+    expect(searchParams.get(key)).to.equal(String(value));
+  });
+  if (params.type === types.checkoutUrl) {
+    expect(searchParams.get('text')).to.equal(text);
+    expect(link.text).to.equal(`CTA {{${text}}}`);
+  } else {
+    expect(link.text).to.equal(`PRICE - ${offer.planType} - ${offer.name}`);
+  }
+}
+
+function createLink(params = {}) {
+  return createLinkMarkup(
+    defaults,
+    params.osi ?? osi,
+    params.type,
+    perpM2M,
+    params,
+    params.promo,
+  );
+}
+
+let origUrl;
+
+describe('OST: loadOstEnv', async () => {
   beforeEach(() => {
     sessionStorage.clear();
+    origUrl = document.location.href;
   });
+
+  afterEach(() => {
+    window.history.replaceState({}, '', origUrl);
+  });
+
   it('fetches and returns page status and metadata', async () => {
     const {
       options: { country, language, workflow },
@@ -81,10 +128,16 @@ describe('loadOstEnv', async () => {
   });
 });
 
-describe('init', () => {
+describe('OST: init', () => {
   beforeEach(() => {
     sessionStorage.clear();
+    origUrl = document.location.href;
   });
+
+  afterEach(() => {
+    window.history.replaceState({}, '', origUrl);
+  });
+
   it('opens OST without waiting for IMS if query string includes token', async () => {
     const {
       options: { country, language, workflow },
@@ -176,6 +229,72 @@ describe('init', () => {
       language,
       wcsApiKey: WCS_API_KEY,
       workflow,
+    });
+  });
+});
+
+describe('OST: merch link creation"', () => {
+  describe('checkout-link', () => {
+    const type = types.checkoutUrl;
+
+    it('with default params', async () => {
+      const link = createLink({ type });
+      assertLink(link, perpM2M, { osi, type });
+      expect({ ...link.dataset }).to.eql({});
+    });
+
+    it('with promo and custom text', async () => {
+      const ctaText = texts.try;
+      const link = createLink({ ctaText, promo, type });
+      assertLink(link, perpM2M, { osi, promo, type }, ctaText);
+    });
+
+    it('to UCv2 workflow', async () => {
+      const workflow = CheckoutWorkflow.V2;
+      const workflowStep = CheckoutWorkflowStep.CHECKOUT_EMAIL;
+      const link = createLink({ type, workflow, workflowStep });
+      assertLink(link, perpM2M, { osi, type, workflow, workflowStep });
+    });
+  });
+
+  describe('inline-price', () => {
+    const type = types.price;
+
+    it('with default params', async () => {
+      const link = createLink({ type });
+      assertLink(link, perpM2M, { osi, type });
+    });
+
+    it('with default params from OST', async () => {
+      const link = createLink({
+        type,
+        displayRecurrence: true,
+        displayPerUnit: true,
+        displayTax: false,
+        displayOldPrice: false,
+        forceTaxExclusive: false,
+      });
+      expect(link.href).to.eql('https://milo.adobe.com/tools/ost?osi=cea462e983f649bca2293325c9894bdd&type=price&perp=true');
+    });
+
+    it('with custom options', async () => {
+      const displayRecurrence = true;
+      const displayPerUnit = true;
+      const displayTax = true;
+      const forceTaxExclusive = true;
+      const link = createLink({
+        displayRecurrence,
+        displayPerUnit,
+        displayTax,
+        forceTaxExclusive,
+        type,
+      });
+      assertLink(link, perpM2M, {
+        tax: displayTax,
+        exclusive: forceTaxExclusive,
+        osi,
+        type,
+      });
     });
   });
 });
